+++ src/mips64/simulator-mips64.h
@@ -424,7 +424,7 @@ class Simulator {
   static const size_t stack_size_ = 1 * 1024*1024;
   char* stack_;
   bool pc_modified_;
-  int icount_;
+  int64_t icount_;
   int break_count_;
   EmbeddedVector<char, 128> trace_buf_;
